1. 玩家首次加入的初始硬币数设置为5000 2. 将经验存储数据类型更改为double，显示时保留2位小数 3. 将击杀获得的经验和硬币更改为double并取消取整

+++ src/main/java/com/ixbob/thepit/event/OnPlayerOwnXpModifiedListener.java
@@ -13,10 +13,10 @@ import org.bukkit.event.Listener;
 public class OnPlayerOwnXpModifiedListener implements Listener {
     @EventHandler
     public void onPlayerOwnXpModified(PlayerOwnXpModifiedEvent event) {
-        int originXp;
-        int modifiedXp;
-        int newOwnXp;
-        int nextLevelNeedXp;
+        double originXp;
+        double modifiedXp;
+        double newOwnXp;
+        double nextLevelNeedXp;
         int originLevel;
         int upgradeLevel = 0;
         int newLevel;
@@ -47,7 +47,7 @@ public class OnPlayerOwnXpModifiedListener implements Listener {
         dataBlock.updateScoreboardNextLevelNeedXp();
         Utils.updateDisplayName(player);
     }
-    private void updateDataBlock(PlayerDataBlock dataBlock, int newLevel, int newThisLevelOwnXp, int newNextLevelNeedXp) {
+    private void updateDataBlock(PlayerDataBlock dataBlock, int newLevel, double newThisLevelOwnXp, double newNextLevelNeedXp) {
         dataBlock.setLevel(newLevel);
         dataBlock.setThisLevelOwnXp(newThisLevelOwnXp);
         dataBlock.setNextLevelNeedXp(newNextLevelNeedXp);
